% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal180to360}
\alias{gdal180to360}
\title{Rotate raster from -180/180 to 0/360 degrees}
\usage{
gdal180to360(inpath, outpath, return.raster = FALSE)
}
\arguments{
\item{inpath}{file path of inpath file to change.}

\item{outpath}{file path of output file to generate.}

\item{return.raster}{if TRUE raster will be returned from function call.}
}
\description{
Rotate raster from -180/180 to 0/360 degrees
}
