% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal_180_to_360}
\alias{gdal_180_to_360}
\title{Rotate raster from -180/180 to 0/360 degrees}
\usage{
gdal_180_to_360(inpath, outpath, return_raster = FALSE)
}
\arguments{
\item{inpath}{file path of input file to change.}

\item{outpath}{file path of output file to generate.}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Rotate raster from -180/180 to 0/360 degrees
}
