% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal_crop}
\alias{gdal_crop}
\title{Crop function using gdal}
\usage{
gdal_crop(inpath, outpath, extent, resolution, return_raster = FALSE)
}
\arguments{
\item{inpath}{file path of the input raster}

\item{outpath}{file path of the output raster}

\item{extent}{extent for raster to be cropped to}

\item{resolution}{resolution of output raster (usually based on the input raster)}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Crop function using gdal
}
