% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal_distance}
\alias{gdal_distance}
\title{Calculate distance raster using gdal}
\usage{
gdal_distance(inpath, outpath, target = NULL, maxdist = NULL,
  return_raster = FALSE)
}
\arguments{
\item{inpath}{file path of input file to change.}

\item{outpath}{file path of output file to generate.}

\item{target}{numeric vaules to calculate distance from default is zero.}

\item{maxdist}{The maximum distance to be generated. maxdist is in pixels.}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Calculate distance raster using gdal
}
