% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_functions.R
\name{gdal_crop}
\alias{gdal_crop}
\title{gdal functions in R}
\usage{
gdal_crop(inpath, outpath, extent, resolution)
}
\arguments{
\item{inpath}{file path of the input raster}

\item{outpath}{file path of the output raster}

\item{extent}{extent for raster to be cropped to}

\item{resolution}{resolution of output raster (usually based on the input raster)}
}
