% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal_mask}
\alias{gdal_mask}
\title{Mask function using gdal}
\usage{
gdal_mask(inpath, mask, outpath, return_raster = FALSE)
}
\arguments{
\item{inpath}{file path of the input raster}

\item{mask}{path to a mask file which is NA for cells to mask and 1 for cells to keep.}

\item{outpath}{file path of the output raster}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Mask function using gdal
}
