% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalCrop}
\alias{gdalCrop}
\title{Crop function using gdal}
\usage{
gdalCrop(inpath, outpath, extent = NULL, resolution = NULL, return = TRUE)
}
\arguments{
\item{inpath}{file path of the inpath raster}

\item{outpath}{file path of the output raster}

\item{extent}{extent for raster to be cropped to}

\item{resolution}{resolution of output raster (usually based on the inpath raster)}

\item{return.raster}{if TRUE raster will be returned from function call.}
}
\description{
Crop function using gdal
}
