% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalEdit}
\alias{gdalEdit}
\title{Mask function using gdal}
\usage{
gdalEdit(inpath, scale, offset, return.raster = TRUE)
}
\arguments{
\item{inpath}{file path of the inpath raster}

\item{scale}{path to a mask file which is NA for cells to mask and 1 for cells to keep.}

\item{offset}{file path of the output raster}

\item{return.raster}{if TRUE raster will be returned from function call.}
}
\description{
Mask function using gdal
}
