% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalExtract}
\alias{gdalExtract}
\title{Extract raster values from points.}
\usage{
gdalExtract(inrasters, pts, simplify = TRUE, sp = FALSE)
}
\arguments{
\item{inrasters}{file path(s) of rasters to extract values}

\item{pts}{Coordinates of points to extract raster values.}

\item{simplify}{Logical. If \code{TRUE}, then if all files referred to in
\code{srcfile} are single band rasters, the sampled values will be
simplified to a matrix.}

\item{sp}{Logical. If \code{TRUE}, a \code{SpatialPointsDataFrame} will be
returned, otherwise a \code{data.frame} will be returned.}
}
\description{
Extract raster values from points.
}
