% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalMask}
\alias{gdalMask}
\title{Mask function using gdal}
\usage{
gdalMask(inpath, mask, outpath, quiet = TRUE, return.raster = FALSE)
}
\arguments{
\item{inpath}{file path of the inpath raster}

\item{mask}{path to a mask file which is NA for cells to mask and 1 for cells to keep.}

\item{outpath}{file path of the output raster}

\item{return.raster}{if TRUE raster will be returned from function call.}
}
\description{
Mask function using gdal
}
