% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalMultiband2Singles}
\alias{gdalMultiband2Singles}
\title{Split a multiband raster into multiple single files.}
\usage{
gdalMultiband2Singles(inpath, outdir = NULL, bands = NULL, return_list = TRUE)
}
\arguments{
\item{inpath}{input multiband raster file path.}

\item{outdir}{output directory to write multiple single files.}

\item{return_list}{TRUE, return rasters as a list.}

\item{band}{which bands to convert, default is NULL (and will do all).}
}
\description{
Split a multiband raster into multiple single files.
}
