% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalProject}
\alias{gdalProject}
\title{Faster reprojectiom of rasters using gdal}
\usage{
gdalProject(
  inraster,
  outdir = NULL,
  xres,
  yres = xres,
  s_srs,
  t_srs,
  resampling,
  extent,
  of = "GTiff",
  extension = "tif",
  ot = "Float32",
  returnRaster = TRUE
)
}
\arguments{
\item{inraster}{local or file path of the inpath raster}

\item{outdir}{local shapefile or file path of the output raster}

\item{xres}{resolution of cells}
}
\description{
Faster reprojectiom of rasters using gdal
}
