% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalReclassify}
\alias{gdalReclassify}
\title{Reclassify raster using gdal_calc}
\usage{
gdalReclassify(
  inpath,
  outpath,
  calc_fun,
  overwrite = TRUE,
  return_raster = TRUE
)
}
\arguments{
\item{inpath}{file path of inpath file to change.}

\item{outpath}{file path of output file to generate.}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Reclassify raster using gdal_calc
}
