% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalResample}
\alias{gdalResample}
\title{Resample raster to different resolution}
\usage{
gdalResample(
  inpath,
  outpath = NULL,
  resolution = NULL,
  method = "bilinear",
  bigtif = FALSE,
  return.raster = TRUE
)
}
\arguments{
\item{inpath}{file path of inpath file to change.}

\item{outpath}{file path of output file to generate.}

\item{resolution}{desired new resolution for x and y.}

\item{method}{resample method, default is nearest neighbour which is fast but dodgy. Options are: "near", "bilinear", "cubic", "cubicspline",}

\item{bigtif}{if TRUE deal with a big geotiff slightly differently.}

\item{return.raster}{if TRUE raster will be returned from function call.}
}
\description{
Resample raster to different resolution
}
