% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalStitchTitles}
\alias{gdalStitchTitles}
\title{Stitch together tiles.}
\usage{
gdalStitchTitles(outpath, tiles_path, return_raster = TRUE, large_tif = TRUE)
}
\arguments{
\item{outpath}{file path of output file to generate.}

\item{return_raster}{if TRUE raster will be returned from function call.}

\item{inpath}{file path of inpath file to change.}

\item{target}{numeric values to calculate distance from default is zero.}

\item{maxdist}{The maximum distance to be generated. maxdist is in pixels.}
}
\description{
Stitch together tiles.
}
