% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalTiles}
\alias{gdalTiles}
\title{Calculate distance raster using gdal}
\usage{
gdalTiles(inpath, outdir, nx = 256, ny = nx, multiInput = FALSE)
}
\arguments{
\item{inpath}{file path of inpath file to change.}

\item{outdir}{file path of output file to generate.}

\item{nx}{The number of x direction pixels per tile}

\item{ny}{The number of x direction pixels per tile}
}
\description{
Calculate distance raster using gdal
}
