% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr_polygon_functions.R
\name{mergeOGR}
\alias{mergeOGR}
\title{Merge multple shapefiles}
\usage{
mergeOGR(infiles.list, outfile)
}
\arguments{
\item{outfile}{The path to shapefile you wish to write too}

\item{infile.list}{A list of path files that points to shape files}
}
\description{
Merge multple shapefiles
}
