% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_polygon_functions.R
\name{ogrBBoxClip}
\alias{ogrBBoxClip}
\title{Clip vector file ogr2ogr with a bounding box}
\usage{
ogrBBoxClip(inshp, outshp = NULL, ext, returnShape = TRUE)
}
\arguments{
\item{returnShape}{if TRUE raster will be returned from function call.}

\item{inpath}{file path of the inpath raster}

\item{outpath}{file path of the output raster}

\item{extent}{extent for raster to be cropped to}
}
\description{
Clip vector file ogr2ogr with a bounding box
}
