% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdal_resample}
\alias{gdal_resample}
\alias{gdalResample}
\title{Resample raster to different resolution}
\usage{
gdalResample(
  inraster,
  outpath = NULL,
  resampleRaster = NULL,
  resolution = NULL,
  method = "bilinear",
  bigtif = FALSE,
  returnRaster = TRUE
)
}
\arguments{
\item{inraster}{file path of inpath file to change.}

\item{outpath}{file path of output file to generate.}

\item{resolution}{desired new resolution for x and y.}

\item{method}{resample method, default is nearest neighbour which is fast but dodgy.}

\item{bigtif}{if TRUE deal with a big geotiff slightly differently.}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Resample raster to different resolution
}
