% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalClipWithShape}
\alias{gdalClipWithShape}
\title{Clip raster with a shape file}
\usage{
gdalClipWithShape(
  inrast,
  inshp,
  outrast = NULL,
  field = NULL,
  cropToShape = FALSE,
  return.raster = TRUE
)
}
\arguments{
\item{inrast}{file path or raster object}

\item{inshp}{file path or shape object}

\item{outrast}{NULL or file path. Default NULL and produces a tmp raster file.}

\item{field}{name of specific field in shapefile to clip raster with. Default is NULL}

\item{cropToShape}{logical Should the raster be cropped to the extent of the shapefile. Default is FALSE}

\item{return.raster}{logical Should the raster be returned to R environment? Default is TRUE}
}
\description{
Clip raster with a shape file
}
