% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalRasterise}
\alias{gdalRasterise}
\title{Faster rasterisation of shapefile to raster using gdal}
\usage{
gdalRasterise(shp, rast, variable = NULL, bigtif = FALSE)
}
\arguments{
\item{shp}{local shapefile or file path of the inpath raster}

\item{rast}{local shapefile or file path of the output raster}

\item{variable}{what variable to convert to a raster}
}
\description{
Faster rasterisation of shapefile to raster using gdal
}
