% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalReclassify}
\alias{gdalReclassify}
\title{Reclassify raster using gdal_calc}
\usage{
gdalReclassify(
  inpath,
  outpath,
  reclassify_list = NULL,
  calc_fun = NULL,
  overwrite = TRUE,
  return_raster = TRUE
)
}
\arguments{
\item{inpath}{file path of inpath file to change.}

\item{outpath}{file path of output file to generate.}

\item{reclassify_list}{A list of values to reclassify, the value will be assigned by list index, e.g the first element will be made 1 and so forth. Values associated with that list will be converted to that number.}

\item{calc_fun}{A text function that follows the "https://gdal.org/programs/gdal_calc.html" syntax.}

\item{return_raster}{if TRUE raster will be returned from function call.}
}
\description{
Reclassify raster using gdal_calc
}
