% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal_raster_functions.R
\name{gdalStitchTitles}
\alias{gdalStitchTitles}
\title{Stitch together tiles.}
\usage{
gdalStitchTitles(tilespath, outpath, bigtif = TRUE, return.raster = TRUE)
}
\arguments{
\item{tilespath}{file path to the raster tiles. use "/path/*.tif"}

\item{outpath}{file path of output file to generate.}

\item{bigtif}{Is the raster very large > 4GB; If so use this call.}

\item{return.raster}{default TRUE, raster returned to R environment.}
}
\description{
Stitch together tiles.
}
